requirements.txt
easyocr==1.7.1
PyPDF2==3.0.1
torch>=2.0.0
numpy>=1.21.0
Pillow>=9.0.0
pdf2image>=1.16.3
opencv-python>=4.8.0
streamlit>=1.24.0
plotly>=5.13.0
pandas>=1.5.0
python-dotenv>=0.19.0
paddleocr>=2.6.0
PyMuPDF>=1.23.0
langchain>=0.0.300
openai>=1.0.0 